requirements.txt
borax==1.0.2
Django==2.0.5
django-cors-headers==2.4.0
django-echarts==0.3.5
django-filter==2.0.0
django-guardian==1.4.9
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
dukpy==0.2.0
et-xmlfile==1.0.1
future==0.16.0
javascripthon==0.9
jdcal==1.4
Jinja2==2.10
jupyter-echarts-pypkg==0.1.0
lml==0.0.2
macropy3==1.0.4.dev2
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
openpyxl==2.5.4
Pillow==5.1.0
pyecharts==0.4.1
pyecharts-javascripthon==0.0.5
pyecharts-jupyter-installer==0.0.3
PyJWT==1.6.4
pytz==2018.4
six==1.11.0
xlrd==1.1.0
